Move everything related to write/to operators to one file write_to.cpp
